Add config options to use po_diff/pot_diff for po files
po file changes in "git log -p" are pretty ugly, so
this set's a "textconv" for "po" file types, and
uses that for a less noisy version of po file
diffs.

See 'gitattributes' and gitconfig for examples
of the setup to turn this on. Need to
add the 'po'/'pot' attr defs to your projects
.gitattributes